Implements goto prev/next page feature
For now I implemented a very naive algorithm that looks for <a> with
matching text. It would probably be better to somehow reuse the logic in
the hint link discovery, but it is actually subtly different here.  In
hint link discovery, we do a DFS from the root document, and stop
recursing whenever we identify a link. But here we actually want to
visit every child, since the text could potentially be on one of the
child elements instead.

Text matching patterns for next/previous page links can be configured as
`nextpagetextpatterns` and `prevpagetextpatterns` options in sVimrc.

Default settings tested on Google, and GitHub issue page.